test: split source-checkout asset tests into test_host_setup_source_assets
The T05 validation command expects tests/unit/service/test_host_setup_source_assets.py,
but the source_assets tests were bundled inside test_host_setup_config.py. Move them
into a dedicated file mirroring src/awf/host_setup/source_assets.py and drop the
now-unused imports from the config test module. No test behavior changes; coverage
is preserved.
